refactor: (#54) 사원 번호 찾기 Query 추가
